SMSAero.API

Documentation

data SmsAeroJson

Content type for SMSAero JSON answer (it has JSON body but "text/plain" Content-Type).

data RequiredQueryParam sym a

Like QueryParam, but always required.

newtype Signature

SMSAero sender's signature. This is used for the "from" field.

Constructors

Signature

Fields

getSignature :: Text

newtype MessageId

SMSAero sent message id.

Constructors

MessageId Integer

data SMSAeroAuth

SMSAero authentication data.

Constructors

SMSAeroAuth

Fields

authUser :: Text

Username.

authPassword :: Text

MD5 hash of a password.

newtype Phone

Phone number.

Constructors

Phone

Fields

getPhone :: Integer

newtype SMSAeroDate

Date.

Constructors

SMSAeroDate

Fields

getSMSAeroDate :: UTCTime

data RequireAuth

SMSAero authentication credentials.

data AnswerJson

Implicit parameter that tells SMSAero to respond with JSON.

type SmsAeroGet a = Get `[SmsAeroJson]` (SmsAeroResponse a)

Regular SMSAero GET API.

type SMSAeroAPI = RequireAuth :> (AnswerJson :> (("send" :> SendApi) :<|> (("status" :> StatusApi) :<|> (("balance" :> SmsAeroGet BalanceResponse) :<|> (("senders" :> SmsAeroGet SendersResponse) :<|> ("sign" :> SmsAeroGet SignResponse))))))

SMSAero API.

type SendApi = RequiredQueryParam "to" Phone :> (RequiredQueryParam "text" Text :> (RequiredQueryParam "from" Signature :> (QueryParam "date" SMSAeroDate :> SmsAeroGet SendResponse)))

SMSAero API to send a message.

type StatusApi = RequiredQueryParam "id" MessageId :> SmsAeroGet StatusResponse

SMSAero API to check message status.

data SmsAeroResponse a

Every SMSAero response is either rejected or provides some info.

Constructors

ResponseOK a
Some useful payload.

ResponseReject Text
Rejection reason.

data SendResponse

SMSAero response to a send request.

Constructors

SendAccepted MessageId
Message accepted.

SendNoCredits
No credits to send a message.

data StatusResponse

SMSAero response to a status request.

Constructors

StatusDeliverySuccess
Message is successfully delivered.

StatusDeliveryFailure
Message delivery has failed.

StatusSmscSubmit
Message submitted to SMSC.

StatusSmscReject
Message rejected by SMSC.

StatusQueue
Message queued.

StatusWaitStatus
Wait for message status.

newtype BalanceResponse

SMSAero response to a balance request. This is a number of available messages to send.

Constructors

BalanceResponse Double

newtype SendersResponse

SMSAero response to a senders request. This is just a list of available signatures.

Constructors

SendersResponse [Signature]

data SignResponse

SMSAero response to a sign request.

Constructors

SignApproved
Signature is approved.

SignRejected
Signature is rejected.

SignPending
Signature is pending.